% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_evs.R
\name{assign_evs}
\alias{assign_evs}
\title{Run EV assignment model}
\usage{
assign_evs(prepped_data, evs_to_add, model_coefficients, prediction_colname)
}
\arguments{
\item{prepped_data}{Dataset prepped for model use (all continuous variables
centered and scaled), probably produced by \code{\link{prep_bg_data}}}

\item{evs_to_add}{Number of additional EVs to add}

\item{model_coefficients}{A data frame of coefficients to use, with one
column of variable names and one column of values ... Current choices
\code{CA_bev} and \code{CA_phev} are stored in
\code{\link{assignment_model_coefficients}}}

\item{prediction_colname}{Name for the new column of predicted PEVs}
}
\value{
A dataframe identical to prepped data with an additional column added
for number of predicted PEVs
}
\description{
Assign EVs to block groups with baseline adoption rates computed using
coefficients included in this package. Typically will be run AFTER calling
\code{\link{prep_bg_data}}
}
\examples{
# First, prepare the input data for this functino
prepped_data <- prep_bg_data()

# Then, run a scenario where 20,000 BEVs and 10,000 PHEVs are assigned:
prepped_data \%>\%
  assign_evs(20000, evworkplace::assignment_model_coefficients$CA_bev,  "bevs_20k") \%>\%
  assign_evs(10000, evworkplace::assignment_model_coefficients$CA_phev, "phevs_10k")
}
