% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_workcharge.R
\name{assign_workcharge}
\alias{assign_workcharge}
\title{Create work charging events and demand}
\usage{
assign_workcharge(
  prepped_data,
  commute_ev_col,
  charging_probs,
  scenario_name,
  extra_daily_miles = 10,
  miles_per_kWh = 4
)
}
\arguments{
\item{prepped_data}{EV commute counts produced by \code{\link{prep_workcharge}}}

\item{commute_ev_col}{Unquoted name of column in \code{prepped_data} that contains the EV count for this scenario}

\item{charging_probs}{Work charging probabilities produced by \code{\link{make_workcharge_probs}}}

\item{scenario_name}{Text name of the combined adoption-charging scenario, used as prefix for column names in the output}

\item{extra_daily_miles}{Non-commute travel per day, miles}

\item{miles_per_kWh}{Vehicle miles per kilowatt hour}
}
\value{
A data frame of work charging events and kWh for the scenario in each block group
}
\description{
Create work charging events and demand for a given EV adoption and charging scenario
}
\examples{
# First, create a work charging dataset using only 50\% of EVs for commuting:
ev_commuter_count <- ev_totals_sample \%>\%
  prep_workcharge(bevs_20k, phevs_10k, .ev_commute_frac = 0.5)

# create a charging scenario each for BEVs and PHEVs
bev_chargeprob  <- make_workcharge_probs(use_BEV = TRUE)
phev_chargeprob <- make_workcharge_probs(use_BEV = FALSE)

# run the workplace charging calculator
ev_workplace_demand_bev20 <- ev_commuter_count \%>\%
  assign_workcharge(bevs_20k, bev_chargeprob, "bev20k_basic")

ev_workplace_demand_phev10 <- ev_commuter_count \%>\%
  assign_workcharge(phevs_10k, phev_chargeprob, "phev10k_basic")
}
