% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assignment_model_coefficients}
\alias{assignment_model_coefficients}
\title{Adoption model coefficients for this study.}
\format{
A named of two data frames with model coefficients, with entries named \code{BEV} and \code{PHEV}. Each data frame has 8 rows and 2 variables:
\describe{
\item{Variable}{Variable name, all but intercept, p, and q must match the name of a column in the block group dataset}
\item{Coefficient}{Coefficient of this column in model for base adoption rate}
}
}
\usage{
assignment_model_coefficients
}
\description{
A dataset containing two sets of model coefficients to use for EV assignment.
}
\details{
Currently included coefficients are derived from models for BEV and PHEV adoption in California.
A paper from this study is in review in \emph{Transportation Research Record} as
"Investigating the sensitivity of electric vehicle out-of-home charging demand to changes in light-duty vehicle fleet makeup and usage, a case study for California 2030"
}
\keyword{datasets}
