% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bp_simplified.R
\name{calculate_bp_simplified}
\alias{calculate_bp_simplified}
\title{Calculate base penetration rate}
\usage{
calculate_bp_simplified(
  fixed_predictor,
  other_predictors,
  coefficients,
  frame = 10
)
}
\arguments{
\item{fixed_predictor}{Vector of values to be used as the fixed predictors}

\item{other_predictors}{Data frame of values to that match names in
coefficients}

\item{coefficients}{Named vector of coefficients, probably derived from (
\code{\link{assignment_model_coefficients}})}

\item{frame}{numeric; maximum divergence allowed for calculated BP. By
default, the smallest BP will be 0.1\*mean, and the largest will be
10\*mean)}
}
\value{
Vector of bp values
}
\description{
This is basically a version of \code{\link[stocked]{calculate_bp}} that works
inside mutate
}
