% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dvrpc_od}
\alias{dvrpc_od}
\title{Commute OD matrix for the DVRPC region}
\format{
A data frame with 1,879,017 rows and 4 variables: \describe{
\item{Home_BlkGrp}{FIPS code for home/origin block group}
\item{Work_BlkGrp}{FIPS code for work/destination block group}
\item{Distance}{Distance between home and work block group, in miles}
\item{total_Jobs}{Total jobs in \code{Work_BlkGrp} held by people who live in \code{Home_BlkGrp}}}
}
\source{
Derived from the US Census LODES dataset.
}
\usage{
dvrpc_od
}
\description{
A dataset containing a commute OD matrix for the Delaware Valley region.
}
\keyword{datasets}
