% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evworkplace-package.R
\docType{package}
\name{evworkplace-package}
\alias{evworkplace}
\alias{evworkplace-package}
\title{evworkplace: Produce Workplace Charging Demand for an EV Ownership Scenario}
\description{
This package estimate workplace charging demand for an EV adoption scenario produced using the stocked package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davisadamw/evworkplace}
  \item Report bugs at \url{https://github.com/davisadamw/evworkplace/issues}
}

}
\author{
\strong{Maintainer}: Adam Davis \email{davisadamw@gmail.com} (\href{https://orcid.org/0000-0001-6324-6380}{ORCID})

}
\keyword{internal}
