% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_workcharge.R
\name{prep_workcharge}
\alias{prep_workcharge}
\title{Assign EVs to commute routes}
\usage{
prep_workcharge(ev_totals, ..., .ev_commute_frac = 0.6)
}
\arguments{
\item{ev_totals}{Dataset containing home block groups and one or more columns of EV counts.}

\item{...}{Unquoted names of variables in \code{ev_totals} containing number of EVs in a block group}

\item{.ev_commute_frac}{What fraction of EVs are used for commuting? Default = 60\%}
}
\value{
A dataset with one row for each home-work pair of block groups in the study area.
}
\description{
Assign EVs to commute routes
}
\examples{
# Take an output from the assign_evs() function and assign to commute destinations:
ev_totals_sample \%>\%
  prep_workcharge(bevs_20k, phevs_10k)

# Run it again, using only 50\% of EVs for commuting:
ev_totals_sample \%>\%
  prep_workcharge(bevs_20k, phevs_10k, .ev_commute_frac = 0.5)
}
